requirements.txt
fastapi>=0.111
uvicorn[standard]
streamlit
mlflow
apache-airflow
prometheus-client
folium
streamlit-folium
ultralytics
torch
torchvision
pytest
httpx
python-multipart